[Minor] Types for Header component
