+++ docs/src/pages/components/cdk-components/header/api.jsx
@@ -19,7 +19,7 @@ const headerPropsTable = () => {
         </td>
       </tr>
       <tr>
-        <td>content: object</td>
+        <td>content: node</td>
         <td></td>
         <td>
           Content showed in the header. Take into account that the component
